feat: URL loading supports both JSON and parquet files for all data arrays
