+++ src/colorizer/index.ts
@@ -2,13 +2,22 @@ import ColorizeCanvas from "./ColorizeCanvas";
 import ColorRamp, { ColorRampType } from "./ColorRamp";
 import Dataset from "./Dataset";
 import ImageFrameLoader from "./loaders/ImageFrameLoader";
-import JsonArrayLoader from "./loaders/JsonArrayLoader";
+import UrlArrayLoader from "./loaders/JsonArrayLoader";
 import Plotting from "./Plotting";
 import Track from "./Track";
 
 export type { ArraySource, IArrayLoader, IFrameLoader } from "./loaders/ILoader";
 
-export { ColorizeCanvas, ColorRamp, ColorRampType, Dataset, ImageFrameLoader, JsonArrayLoader, Plotting, Track };
+export {
+  ColorizeCanvas,
+  ColorRamp,
+  ColorRampType,
+  Dataset,
+  ImageFrameLoader,
+  UrlArrayLoader as JsonArrayLoader,
+  Plotting,
+  Track,
+};
 
 export * from "./colors/categorical_palettes";
 export * from "./colors/color_ramps";
