+++ src/colorizer/utils/data_utils.ts
@@ -5,7 +5,8 @@ import { FeatureThreshold, isThresholdCategorical, isThresholdNumeric, Threshold
 import ColorRamp from "../ColorRamp";
 import Dataset, { FeatureType } from "../Dataset";
 
-/** Replaces all NaN in string text (such as the string representation of a JSON
+/**
+ * Replaces all NaN in string text (such as the string representation of a JSON
  * object) with null. Can be used to safely parse JSON objects with NaN values.
  */
 export const nanToNull = (json: string): string => json.replace(/NaN/g, "null");
